Moved ConvertPulsesToMillimeters for bays from BaysDataProvider to InverterProvider + Fixed mixed usage of double/decimal for Resolution fields (now it's double)

+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Data/ResolutionCalibrationMessageData.cs
@@ -25,7 +25,7 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 
         public double ReadStartPosition { get; }
 
-        public decimal Resolution { get; set; }
+        public double Resolution { get; set; }
 
         public MessageVerbosity Verbosity { get; }
